Move LabelForm into LabelPage directory
