+++ frontend/pages/LabelPage/LabelForm/LabelForm.tsx
@@ -12,7 +12,7 @@ import InputField from "components/forms/fields/InputField";
 import FleetAce from "components/FleetAce";
 // @ts-ignore
 import validateQuery from "components/forms/validators/validate_query";
-import InfoIcon from "../../../../../../assets/images/icon-info-purple-14x14@2x.png";
+import InfoIcon from "../../../../assets/images/icon-info-purple-14x14@2x.png";
 
 interface ILabelFormProps {
   baseError: string;
